refactor: Make the GBTS graph representation internal
Nothing outside `Core/src` names `GbtsNodeParams`, `GbtsNodeEdgeInfo`,
`GbtsEtaBinInfo`, `GbtsNodeView`, `GbtsNodeProxy`, `GbtsEdge` or
`GbtsEdgeState`. They move to `Acts::Experimental::detail`, the graph types into
a new `Acts/Seeding/detail/GbtsGraphTypes.hpp` and the filter state into
`Acts/Seeding/detail/GbtsEdgeState.hpp`.

`GbtsNodeIndex` stays public - `GbtsNodeStorage::spacePointIndex` takes one - and
moves to a new `Acts/Seeding/GbtsTypes.hpp` that both the storage and the graph
types include.

The accessors that hand them out follow. `GbtsNodeStorage::etaBin`, `nodeView`,
`nodeParams` and `nodeEdgeInfo`, and `GbtsTrackingFilter::followTrack` with its
`State` and `kGbtsMaxEdgeStates`, become private with `GraphBasedTrackSeeder` as
a friend. The seeder's own `SeedCandidateProperties`, `OutputSeedProperties` and
`SlidingWindow` move to its private section.

What stays public is what an experiment actually calls: the seeder config,
`makeNodeStorage` and `createSeeds`, `GbtsNodeStorage::insert`, `extend`,
`finalize`, `numberOfNodes` and `spacePointIndex`, and the tracking filter's
config and constructor.
